FIX: Twitter login problem check not reporting accurately (#35593)
This commit fixes a bug where we were using the twitter's oauth2
endpoint to check if the `twitter_consumer_key` and
`twitter_consumer_secret` has been configured
correctly. However, the request returns a 403 response code because of a
missing required param causing the problem check to fail.

Instead of fixing the request, the better thing to do is to ensure we do
the same checks that the `omniauth-oauth` gem does so that we are
actually validating the oauth flow.