FIX: improves modal on ios when input is focused
This commit mostly various hacks in place and replace them by using the new modifier: `{{preventScrollOnFocus}}`.

Applied this fix to:
- emoji picker modal
- fast edit modal
- add post notice

We might have other modals to fix in the future. Generally the fix should be the same, order matters though, and the modifier has to be before any `{{autoFocus}}` or similar behavior.