FIX: ensures edited post in fast edit is not mutated
The `toggleFastEdit` has an async function in the middle of it and we were using a global post object which can change during the request (most likely be nulled) which can lead to unexpected behavior. The edited post is now saved at the beginning of the toggling preventing this issue.