FIX: improves modal on iOS when input is focused (#35619)
This commit removes various hacks and replace them by using the new
modifier: `{{preventScrollOnFocus}}`.

Applied this fix to:
- emoji picker modal
- fast edit modal
- add post notice

We might have other modals to fix in the future. Generally the fix
should be the same, order matters though, and the modifier has to be
before any `{{autoFocus}}` or similar behavior.