FIX: ensures edited post in fast edit is not mutated (#35622)
The `toggleFastEdit` has an async function in the middle of it and we
were using a global post object which can take time and things can
happen to our post object (most likely be nulled) which can lead to
unexpected behavior. The edited post is now saved at the beginning of
the toggling preventing this issue.

There are already specs for this behavior, but this specific case is
hard to test.