DEV: Modernize composer-messages implementation (#35627)
Replaces resolver-based template lookups with simple components.

Added a value transformer for plugins to modify the components if
needed. This does not appear to be a common or documented pattern, so
this change does not include backwards-compatibility for plugins which
may have been adding their own messages in the `templates` directory