DEV: Refactor email checks to use error classes and more specific error messages
The previous code did not include enough useful information that would enable
the user to make a report asking for assistance. We now include the actual
error class and message in the output.

Formatting of the actual output blocks now resizes to the user's terminal width
and is centrally handled.

Reworded the error messages for consistency and to offer additional advice.

Each of these scenarios outlined in the commit was tested using a corresponding
local setup to generate that error.