FIX: (safari-only) avoid image duplication on cross-window drop to rich editor (#35643)
Exclusively on Mac Safari, dragging and dropping an image from another
window to the composer is processing the image twice, one as an HTML
paste, another as an upload.

This `drop` handler is intended to delegate the processing so Uppy's
`DropTarget` can handle the upload.

This is very hard to add a test for, unfortunately.