FIX: avoid ProseMirror handling paste when unauthorized (#35640)
When `authorizesOneOrMoreExtensions` is `false`, we don't call
`ComposerUpload#setup` to add the paste event listener, which originally
handles the `preventDefault`:

https://github.com/discourse/discourse/blob/d1e1c02fcb0cbc6722280afb6e086cb6a8b53882/frontend/discourse/app/components/composer-editor.gjs#L274-L276

https://github.com/discourse/discourse/blob/d1e1c02fcb0cbc6722280afb6e086cb6a8b53882/frontend/discourse/app/lib/uppy/composer-upload.js#L151-L153

https://github.com/discourse/discourse/blob/d1e1c02fcb0cbc6722280afb6e086cb6a8b53882/frontend/discourse/app/lib/uppy/composer-upload.js#L566-L569

This makes it so ProseMirror handles paste events itself, which leads to
base64 data src images. This PR adds a custom paste handler for
`prosemirror-editor` that prevents default in this case.