FIX: Do not show toast when saving a draft with no changes
When editing a draft, if you open the composer, make no changes and
press X we still save the draft which is a bit pointless. We should only
save if something changes.

This commit also fixes the draft fabricator, which didn't store the
data or the topic sequence key correctly.