DEV: Allow contracts to inherit from another class (#35599)
This change aims to allow contract reusability. Typically when two
services are closely related (like `create`/`update`), a lot of (maybe
even all of them) parameters could be the same.

So instead of having to redefine virtually the same contract twice, with
this change it’s possible to provide a base class for the contract, like
this:
```ruby
params base_class: MyCreationService::Contract
```