FIX: Display a thumbnail for Youtube videos with the classic onebox
Currently, a thumbnail is usually generated when generating oneboxes for
Youtube videos. This happens because of the lazy-videos plugins.
But when it’s not enabled or it can’t fetch the metadata as expected, we
fallback on the original Youtube onebox.

That onebox just outputs an iframe and nothing more. This means no
thumbnails will be created for the associated topic.

This patch addresses the issue by outputting an invisible image before
the iframe, thus allowing thumbnails to be created. This makes the
user experience more consistent, whether using the classic onebox or the
lazy-videos one.