DEV: Don’t ever dump the DB schema (#35736)
We don’t want to dump the DB schema as it can lead to various problems,
mainly with plugins.

We already have some patches that take care of that, but sometimes a
`db/structure.sql` file is still generated.

The solution is to instruct Rails to never dump that schema, even in the
test and development environments.