FIX: Don't require restart for first flag badge working with custom flag (#35733)
When a user uses a newly created flag to flag a post for the first time, they aren't awarded the `First Flag` badge.

The query for the badge includes the IDs of the eligible flags. These are read once on boot and assigned to a constant. So newly created flags aren't eligible until after a restart.

This fixes that by:

- Adding (initially failing) regression tests for the badge query.
- Porting the Ruby method generating the IDs to make the query pure SQL. This ensures it always operates on the relevant data.