PERF: Speed up `TopicQuery#filter_by_tags` with `match_all_tags`
Before this commit, `TopicQuery#filter_by_tags` with the `match_all_tags` option will added an `INNER JOIN topic_tags tx ON tx.topic_id = topics.id AND tx.tag_id = <:tag_id>` for each tag
that we intend to match on. This means that for each row in the `topics`
table that the query scans, it needs to scan the `topic_tags` table once
for each tag. This is very inefficient and does not scale well.

Instead of joining on the `topic_tags` table, we can instead alter the
query to do one `topic_tags` table lookup per tag by using a
subquery. Only a single scan per tag is required for the entire query.