DEV: lint more files using lefthook
lefthook was not linting Gemfile or various files in bin/ which are all ruby files

This will ensure they are linted.

One caveat is if we place non Ruby executables in bin/ we may need to make the
filter more deliberate.

Additionally fixes glint linting (which sadly is a bit slow