FIX: Hashtag icon for style_type=icon not showing in rich editor (#35786)
When selecting a chat channel or a category with style_type=icon
from the hashtag autocomplete in the rich editor, the icon
was not being displayed correctly.

This was happening because when we reused the cache for hashtags
in a61009284e7cc6e237afbb404424f3b67cb0f42b, it exposed an earlier
issue made in d06c60ca7c5081c8080338645910cd7f17504d2e where we
override the `icon` property of the hashtag in the response
with the icon HTML.

Then, in the prosemirror hashtag extension, we would call
`hashtagTypeClass.generateIconHTML(validHashtag)`, which would
then do something like `iconHTML(hashtag.icon)`. But since
`hashtag.icon` was already the HTML, it would produce invalid HTML
similar to `<span class="d-icon-<span
class="d-icon-chat"></span>"></span>`.

This issue did not happen for style_type=emoji or style_type=square
because they had their own manual rendering methods.

The fix was to avoid overriding the `icon` property, instead storing
the icon HTML in a new property `iconHTML` in the hashtag object.