FIX: emojis should insert right after punctuation correctly (#35830)
Context: https://meta.discourse.org/t/302931

There's a discrepancy between when the emoji autocomplete is allowed to
trigger and how we decide where to position its autocompleted term once
selected. I've extracted out the regex pattern used for this, and that
constant will be moved in the near future to a EmojiAutocompleteResults
component that better represents this type of autocomplete (atm, the
next closest semantic module is at
https://github.com/discourse/discourse/blob/c7ba17369645e66593a246f577005da5cce23ff1/frontend/discourse/app/lib/autocomplete/emoji.js
which is planned for removal in an upcoming refactor to use a component
instead).