PERF: DirectoryItem.refresh! daily can take an extremely long time
We were seeing pathological performance when running DirectoryItem.refresh!
in some sites, where it could take _up to 1 hour_ to complete, for an
hourly job.

There were a few main issues with the query:

1. We were doing `COALESCE(created_at, :since)` on user_actions, where
   it is not possible to have NULL created_at. This meant that the
   database could not use an index on created_at, and had to do a full
   table scan.
1. We were doing N1 subplans per user when counting user_visits and
   topic_views, and joining on user_visits twice
1. We were filtering users to id > 0 (human users) (and active/silenced)
   multiple times in subqueries, instead of once in a CTE.
1. We replaced LEFT JOINs with EXISTS where appropriate, which can be
   more efficient, since we only care about existence, not the actual
   data.

All these changes led to a ~1h query going down to ~3s on a large production
dataset.