FEATURE: Add a multiplier for max tokens, typically used for various llm settings like reasoning (#35859)
Meta:
https://meta.discourse.org/t/discourse-ai-failing-to-translate-large-number-of-posts-and-topics/385414

This commit adds a new setting which allows users to modify the max tokens
limit via a multiplier. There are good defaults in place already, but in
reasoning models they seem to require more than the defaults. As there
can be many other reasons this max needs to be bumped up, we're
providing a generic multiplier as a setting.