FIX: Prevent translation editor auto discard
## :mag: Overview
When adding/editing translations using the translation editor, the unsaved changes can be automatically discarded without seeing a warning modal if the following actions were taken:

1. Add/edit a translation
2. **minimize composer**
3. Click reply or new topic

Specifically, minimizing the composer, creates a situation where unsaved changes for translations can be automatically discarded.

This update fixes that issue by checking for `anyDirty` before auto closing as well as ensuring the originalTitle and text properties are set when working with translations.