FIX: ensures calendar is full height
A recent commit https://github.com/discourse/discourse/commit/0bd4cea3fe248228c5a3f9b9599ddd88be38d94a has removed an element in the page which was setting the max height of the container for the upcoming events calendar page. This commit is ensuring upcoming events page is setting its size itself to be the exact height of the page.

To make this possible we had to introduce a css var to size the powered by present at the bottom of the page.