FIX: Do not change language for logged in user when tl is present (#35942)
Meta:
https://meta.discourse.org/t/set-locale-from-param-and-content-localisation-issue/387547

We want to make sure that the `SiteSetting.set_locale_from_param`
conforms to its description: `"Allows setting an anonymous user's locale
via the 'tl' URL param, e.g. ?tl=es"`. Currently, the bug is that this
is applied to logged in users as well.

We also do not want to modify a logged in user's locale as it should
remain as per their user profile.