UX: Copy change for consolidated pageviews report (#35958)
Currently the wording admins see when `use_legacy_pageviews`
is true for the combined browser, crawler, and other pageview
report is:

> Consolidated Pageviews with Browser Detection (Experimental)

However it's not really experimental anymore, we moved a lot
of sites over to using BPV, so we can change Experimental
to Recommended here.

Sites with `use_legacy_pageviews` set to true will see
"Site traffic" for this report instead.