FIX: Use oEmbed for YouTube oneboxing
We currently rely on Open Graph data to render YouTube oneboxes.
However, YouTube does not always give us a response with the Open Graph
data necessary, leading to empty preview HTML like this that shows
as a broken image in the composer:

```
<img src="" width="480" height="360" title=" - YouTube" style="aspect-ratio: 480 / 360;">
```

In addition, our old method of parsing YouTube script tags for JSON
which contains information for the video, introduced in
https://github.com/discourse/onebox/commit/4d669d2b710483281df1b99ff1140acdada11286,
seems to no longer work reliably, possibly due to changes in YouTube's
JS structure.

To fix this, we switch to using YouTube's oEmbed endpoint which gives us
all the metadata we need for oneboxes in JSON format, including title, author,
thumbnail URL, and video dimensions. This approach is more robust and less likely to break
due to changes in YouTube's page structure.

c.f. https://meta.discourse.org/t/youtube-uris-fail-to-render-thumbnails-when-oneboxed/387673