UX: Show chat indicator on mobile (#35973)
We overlooked the consequence of removing the chat icon in the header in
#35015 so this commit brings it back, in a different way.
On mobile only, the back button will save space to display the unread
indicator, if needed:

<img width="732" height="1568" alt="CleanShot 2025-11-11 at 15 53 50@2x"
src="https://github.com/user-attachments/assets/6ddff73b-3f5c-493b-ad44-ec130c8be491"
/>

As a consequence, there will be extra white-space between back-button
and channel title: this is intentional, unavoidable (alternative is
layout shift), and in line with dominating message apps.

This commit also consolidates the unread indicator styling a bit more.