add impression tracking to ads frontend and system tests (#35910)
This is frontend changes and system tests to add impression logging to
ads on the ad plugin. If an ad is seen by a user, and the new site
setting is active, the component lifecycle should fire off a POST
request to `/ad_plugin/ad_impressions` to log the entry in the table.
Added in an intersection observer to only load impressions if the ad is
visible in the window