FIX: Do not use image node for non-image upload placeholders in RTE
When a user uploads a non-image file in the rich text editor, the
system should not create an image node as a placeholder. Instead, it should
use a more appropriate representation, like a link. Otherwise we show a
small broken image icon which is misleading.

This commit makes that change, and also handles success/failure/cancellation
of the upload correctly for non-image files, removing the placeholder link.