FIX: Ensure retry_count is set when checking if a job is being retried (#36034)
When checking how many times an email job has been retried, we need to check
that the `retry_count` attribute has been set.

In a related error, the `UserEmail` retry wait handler assumed that the
triggering exception would always have a `wrapped` attribute, which
isn't necessarily the case if the exception was raised outside of the
job execution code.