FIX: Image optimization error when re-encoding to JPEG
Followup 782e0ea394acfc3caf56483c088668c2240bad56

When we refactored the code in the above commit, we missed
the need to wrap the potentially resized image data in the
media-optimization-bundle with `Uint8ClampedArray` before
passing it to `encode` from `@jsquash/jpeg`. This led to
this cryptic error in Chrome:

```
media-optimization-worker.js:24 TypeError: Failed to construct 'ImageData': The provided value is not of type 'ImageDataSettings'.
    at globalThis.optimize (media-optimization-bundle.js:116:5)
    at onmessage (media-optimization-worker.js:7:42)
```

And this much more helpful error in FireFox:

```
TypeError: ImageData constructor: Argument 1 does not implement interface Uint8ClampedArray.
    optimize media-optimization-bundle.js:116
    onmessage media-optimization-worker.js:7
    EventHandlerNonNull* media-optimization-worker.js:1
```

The difference in errors seems to be due to differences in how the browsers
overload the `ImageData` constructor. Anyway, the fix here is to
wrap `maybeResized` in a `Uint8ClampedArray` before passing it to
`encode`, which is what we do before passing it to `resize`.