FIX: emoji picker autocomplete in chat replaces entire draft (#36017)
Fixes the bug reported in:
https://meta.discourse.org/t/emoji-is-inserted-at-the-wrong-place/302931/12

The textarea element's `selectionStart`/`selectionEnd` properties are
reset once the emoji-picker menu opens and changes focus. This results
in the replacement of the entire textarea's text in
https://github.com/discourse/discourse/blob/d1e1c02fcb0cbc6722280afb6e086cb6a8b53882/frontend/discourse/app/lib/textarea-text-manipulation.js#L747.

The solution here records the relevant start/end position of the emoji
term to be replaced by a selected emoji, before opening the
emoji-picker. This allows us to correctly replace the relevant emoji
term within the chat draft.


https://github.com/user-attachments/assets/f0007883-9420-411a-b9af-9231cc0b827c