UX: Improve template filtering with usage count and preserving tags (#35881)
On sites with a large amount of templates, the default display and
filter can become ineffective. By default, templates are ordered by
alphabetical title order. If filtered by keyword, the plugin does a
simple scoring based on whether the text is in the title (+2) or in the
content (+1), then by alphabetical title order. This PR updates it to
use the `usages` count, so the new priority of ordering is by score >
usage > title.

Also adds a localstorage for the selected tag used to filter the
template list. The tag field will be pre-populated with the last used
filtered tag if available.

Renames hint from "Filter by title..." to "Filter by keyword..." to
reflect actual filtering behavior.