DEV: Add a new ReviewableActionLog model. (#36057)
As part of the review queue refresh, moderators will be able to perform
individual actions on each part of a reviewable (for example, hiding a
post, and suspending the user). In order to know when a reviewable is
"completed", we need to be able to check that the right number of
actions have been performed on the reviewable.

This change adds a `ReviewableActionLog` model where we can keep track
of the actions that have been performed on each reviewable.