DEV: Add keep_images option to ExcerptParser
Previously the only options we had for making excerpts
was to completely strip images or to replace them with
the markdown equivalent e.g. ![alt]. This doesn't look
nice in certain scenarios where you want excerpts but
you do want to see images still too.

This commit adds `keep_images: true` option to the
ExcerptParser which leaves images in the excerpt,
and adds extensive tests for the different image options.