More clearly describe effect of reviewable note copy option
The previous description of this option was quite vague. It might be interpreted as resulting in a publicly visible note
being added to the user's profile (as occurs with the "reason" provided in the suspension and silencing dialogs), or
perhaps even as that the note will be sent to the user (this is an alternate meaning of "copy").

The new wording will ensure staff understand the true effect of the setting.