DEV: Gracefully handle video thumbnail errors (#36087)
If there is an error generating or accessing a video thumbnail during
the video upload process don't block the video upload process. This will
allow videos to continue to be uploaded, but they will have a missing
thumbnail image. This is better than not have an upload at all though.

Added `cleanup()` to avoid possible memory leaks.