UX: allows to click on topic item for bulk select (#36102)
Prior to this commit, you had to exactly click on the checkbox to select
a topic row, now you can click anywhere in the row. It should especially
be better on mobile.

On top of this we have also added a background effect for selected rows
and an active state for the bulk select button.

![Kapture 2025-11-19 at 15 43
35](https://github.com/user-attachments/assets/42a9502f-2baf-4c7d-ada8-ac3589715204)