DEV: Convert keyboard-shortcuts lib to service (#36098)
This lib file required an owner, and had a setup/teardown hooks, so it
really makes more sense as a service.

This also allows us to use `@service` injections, which are much better
than eager `.lookup()` calls. This will resolve issues we're seeing with
theme/plugin `modifyClass` calls racing with these service lookups.