UX: Improve AI summary modal animation (#36108)
## :mag: Overview

We recently updated the animation for modals in core:
https://github.com/discourse/discourse/commit/5263c1632738e3f2c2afd4af9e0187924b96c214.
Although this looks great for all modals, the summary modal in Discourse
AI is unique in that it is positioned at the top right of the page
without a backdrop. As such, the typical scaling and rising up modal
isn't suitable here.

To improve this, this PR updates the animation for the AI summary modal
(on desktop) to float in from the right. Rather than contain the
animation styles in Discourse AI, this update contains the changes in
core's stylesheet in `/base/discourse.scss` (along with some additional
animations in the opposite directions) since these could be useful for
other parts of core later on.

## 📹 Preview

### ← Before

https://github.com/user-attachments/assets/ce5fa7d5-55e1-4a87-95e4-1d21d56e94ee

### → After

https://github.com/user-attachments/assets/e836a834-f8de-413b-95f1-09d415908f77