FEATURE: log errors when enabling discourse id
When admins try to enable discourse id via the site setting, the
automated registration process happen in the background. In the case of
an error, nothing was bubbled up back to the admin other than an
unhelpful error message.

On top of that, no errors were logged to help tech-savvy admins to
figure out what might be the problem(s).

This commit ensures that any error that is happening during the
registration process will be logged via the standard Rails logs.

On top of that, the error message shown to the admin has been updated to
hopefully provide a more helpful description.

Ref - https://meta.discourse.org/t/388711